move hnsw_index_manager not to expose.
